feat: named routes for the application
