+++ lib/presentation/widgets/restaurant_is_open_widget.dart
